fix: expense layout
